moving new structure
