--- src/app/card/card.component.html
@@ -1,10 +1,0 @@
-
-<div class="container">
-    <div class="card" *ngFor="let item of characters">
-            <div>
-             <img src="{{item.thumbnail.path}}.{{item.thumbnail.extension}}">
-            <h3 class="responsive">{{item.name}}</h3>
-            <p class="responsive">{{item.modified}}</p>
-            </div>
-    </div>
-</div>
