+++ src/app/components/card-comic/card-comic.component.css
