+++ src/app/components/card-series/card-series.component.spec.ts
