+++ src/app/components/card-series/card-series.component.ts
