+++ src/app/components/card/card.component.css
