+++ src/app/components/card/card.component.ts
@@ -1,5 +1,11 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
 
+export interface ICard{
+  title:string,
+  modified:Date,
+  thumbnailUrl:string,
+  thumbnailExtension:string
+}
 
 
 @Component({
@@ -8,12 +14,12 @@ import { Component, EventEmitter, Input, Output } from '@angular/core';
   styleUrls: ['./card.component.css']
 })
 export class CardComponent  {
-  
+
   constructor() { }
 
-  @Input() characters: any = [];
-  @Input() card : any =[this.characters];
-   
-  
+  @Input() elements: any = [];
+  @Input() card : ICard[] =[this.elements];
+
+
 }
