+++ src/app/components/character-list/character-list.component.css
