+++ src/app/components/characters/characters.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { CharactersService } from '../shared/services/characters.service';
+import { CharactersService } from '../../shared/services/characters.service';
 
 
 @Component({
@@ -10,23 +10,21 @@ import { CharactersService } from '../shared/services/characters.service';
 export class CharactersComponent implements OnInit {
 
   characters: any = [];
-  
+
 
   constructor(private service: CharactersService) { }
-  
+
   ngOnInit(): void {
    this.all();
-   this.CharacterId();
-
   }
-  
+
    all(){
       this.service.getCharacters().subscribe(resp => {
         this.characters = resp;
      })
    }
 
-  
+
 
   CharacterId(){
      this.service.getCharacterById(this.service.characterId)
@@ -34,5 +32,5 @@ export class CharactersComponent implements OnInit {
      this.characters=res;
        });
     }
-  
+
 }
