+++ src/app/components/comic-list/comic-list.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { debounce } from 'lodash';
-import { ComicsService } from '../shared/services/comics.service';
+import { ComicsService } from '../../shared/services/comics.service';
 
 @Component({
   selector: 'app-comic-list',
@@ -8,13 +8,13 @@ import { ComicsService } from '../shared/services/comics.service';
   styleUrls: ['./comic-list.component.css']
 })
 export class ComicListComponent  {
-  
+
   constructor(private service : ComicsService) { }
   value: string="";
  @Input() comics: any = [];
 //  @Input() cardData: any=[this.characters];
  onChange (e:any){
-  
+
   this.comicByName(e)
  }
 
@@ -22,7 +22,7 @@ export class ComicListComponent  {
  debouncedOnChange =debounce(this.onChange, 250, { 'maxWait': 1000 });
 
  comicByName(title:string){
- 
+
   this.service.getComics(title).subscribe(resp => {
    this.comics = resp;
 })
