+++ src/app/components/search/search.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit,EventEmitter,Output,Input, AfterContentInit, ElementR
 import { FormControl } from '@angular/forms';
 import { debounceTime, distinctUntilChanged, filter, fromEvent, map, tap } from 'rxjs';
 import { CharactersComponent } from '../characters/characters.component';
-import { CharactersService } from '../shared/services/characters.service';
+import { CharactersService } from '../../shared/services/characters.service';
 
 @Component({
   selector: 'app-search',
@@ -10,13 +10,13 @@ import { CharactersService } from '../shared/services/characters.service';
   styleUrls: ['./search.component.css']
 })
 export class SearchComponent  {
- 
- 
+
+
   @Input() characters:any=[];
   @Output() charactersValue = new EventEmitter();
   constructor(private service : CharactersService){}
 
- 
+
    onChange (e:any){
     this.charactersByName();
    }
@@ -25,13 +25,13 @@ export class SearchComponent  {
     this.service.getCharacters(this.service.name).subscribe(resp => {
      this.characters = resp;
   })
-  
+
   }
 
 
 
-  
- 
+
+
 }
