+++ src/app/components/series-list/series-list.component.css
