+++ src/app/components/series-list/series-list.component.html
@@ -1,15 +1,13 @@
 <div class="list">
-    
+
 
     <div class="input">
-        <input  type="search" class="input-search" placeholder="Search by name" [(ngModel)]="value" (ngModelChange)="debouncedOnChange($event)" >
+        <input  type="search" class="input-search" placeholder="Search by title" [(ngModel)]="value" (ngModelChange)="debouncedOnChange($event)" >
         <svg xmlns="http://www.w3.org/2000/svg" class="icon" viewBox="0 0 20 20" fill="currentColor">
           <path fill-rule="evenodd" d="M8 4a4 4 0 100 8 4 4 0 000-8zM2 8a6 6 0 1110.89 3.476l4.817 4.817a1 1 0 01-1.414 1.414l-4.816-4.816A6 6 0 012 8z" clip-rule="evenodd" />
         </svg>
       </div>
-      
-<app-card [characters]="characters" *ngFor="let item of characters" >
-  
-</app-card>
+
+<app-card-series [series]="series"></app-card-series>
 
 </div>
