+++ src/app/components/series-list/series-list.component.spec.ts
