+++ src/app/components/series-list/series-list.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { debounce } from 'lodash';
-import { SeriesService } from '../shared/services/series.service';
+import { SeriesService } from '../../shared/services/series.service';
 
 @Component({
   selector: 'app-series-list',
@@ -13,15 +13,16 @@ export class SeriesListComponent {
   constructor(private service : SeriesService) { }
 
   onChange (e:any){
-  
+
     this.seriesByName(e)
    }
-  
-  
+
+
+
    debouncedOnChange =debounce(this.onChange, 250, { 'maxWait': 1000 });
-  
+
    seriesByName(title:string){
-   
+
     this.service.getSeries(title).subscribe(resp => {
      this.series = resp;
   })
