+++ src/app/components/series/series.component.spec.ts
