+++ src/app/shared/services/characters.service.ts
@@ -5,7 +5,7 @@ import { Character } from '../models/characters/Character';
 import { map, } from 'rxjs/operators';
 import { GlobalConstants } from '../models/common/global-constants';
 import { DataSource } from './datasource.service';
-import { SearchComponent } from 'src/app/search/search.component';
+import { SearchComponent } from 'src/app/components/search/search.component';
 
 
 
@@ -18,12 +18,12 @@ export class CharactersService extends DataSource{
   characterId= 1009664;
   name= "";
   pass:boolean=true;
-  
+
   constructor(public http: HttpClient, ) {
-    
+
     super();
   }
-  
+
   public getCharacters(name?:String):Observable<Character[]>{
     let url=`${GlobalConstants.apiUrl}characters?`;
     if(!!name){
@@ -33,11 +33,10 @@ export class CharactersService extends DataSource{
 
     return this.http.get(url).pipe(
       map((res:any)=>{
-        console.log(res.data.results);
         return res.data.results;
       })
     )
-    
+
   }
 
   public getCharacterById(characterId : number):Observable<Character[]>{
@@ -46,7 +45,6 @@ export class CharactersService extends DataSource{
 
   return this.http.get(url).pipe(
     map((res:any)=>{
-      console.log(res.data.results)
       return res.data.results;
     })
   )
